third_party/blink/web_tests/platform/generic/fast/events/page-visibility-unload-expected.txt

CONSOLE ERROR: Blocked alert('visibilitychanged') during visibilitychange.
You should have seen a warning message for alert dialog.
